Blockly integration into backend: split __init__.py into __init__.py (with BackendServer) and BackendCore + BackendBlockly
